CVar for storage icon label drawing
