+++ Content.Client/UserInterface/Systems/Storage/Controls/ItemGridPiece.cs
@@ -1,11 +1,10 @@
 using System.Numerics;
+using Content.Client._RMC14.Storage;
 using Content.Client.Items.Systems;
-using Content.Shared._RMC14.IconLabel;
 using Content.Shared.Item;
 using Content.Shared.Storage;
 using Robust.Client.GameObjects;
 using Robust.Client.Graphics;
-using Robust.Client.ResourceManagement;
 using Robust.Client.UserInterface;
 using Robust.Client.UserInterface.CustomControls;
 
@@ -21,8 +20,6 @@ public sealed class ItemGridPiece : Control, IEntityControl
     public readonly EntityUid Entity;
     public ItemStorageLocation Location;
     public ItemGridPieceMarks? Marked;
-    // RMC14 - Declare Font for the icon label
-    private Font _font;
 
     public event Action<GUIBoundKeyEventArgs, ItemGridPiece>? OnPiecePressed;
     public event Action<GUIBoundKeyEventArgs, ItemGridPiece>? OnPieceUnpressed;
@@ -55,10 +52,6 @@ public sealed class ItemGridPiece : Control, IEntityControl
     public ItemGridPiece(Entity<ItemComponent> entity, ItemStorageLocation location,  IEntityManager entityManager)
     {
         IoCManager.InjectDependencies(this);
-
-        // RMC14 - Prepare Font for the icon label
-        var cache = IoCManager.Resolve<IResourceCache>();
-        _font = new VectorFont(cache.GetResource<FontResource>("/Fonts/NotoSans/NotoSans-Regular.ttf"), 8);
 
         _entityManager = entityManager;
         _storageController = UserInterfaceManager.GetUIController<StorageUIController>();
@@ -221,36 +214,7 @@ public sealed class ItemGridPiece : Control, IEntityControl
             }
         }
 
-        // RMC14 - Draw text for the icon label (disabled until Giant Text bug is fixed)
-        /**
-        if (_entityManager.TryGetComponent(Entity, out IconLabelComponent? iconLabel))
-        {
-            var scale = 2 * UIScale;
-            if (iconLabel.LabelTextLocId is null ||
-                !Loc.TryGetString(iconLabel.LabelTextLocId, out string? msg))
-            {
-                return;
-            }
-
-            var textColor = Color.Black;
-            Color.TryFromName(iconLabel.TextColor, out textColor);
-
-
-            var charArray = msg.ToCharArray();
-            var iconLabelPosition = new Vector2(iconPosition.X + scale * iconLabel.StoredOffset.X 
-                , iconPosition.Y + scale * iconLabel.StoredOffset.Y);
-
-            var textSize = iconLabel.TextSize;
-
-            float sep = 0;
-            foreach (var chr in charArray)
-            {
-                iconLabelPosition.X += sep;
-                sep = _font.DrawChar(handle, new System.Text.Rune(chr), iconLabelPosition, textSize * scale, textColor);
-            }
-        }
-        **/
-        // RMC14 - End Draw text for the icon label
+        _entityManager.System<RMCIconLabelsSystem>().DrawStorage(Entity, UIScale, iconPosition, handle);
     }
 
     protected override bool HasPoint(Vector2 point)
